% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_flight_path.R
\name{create_flight_path}
\alias{create_flight_path}
\title{Create a smart AGL flight path}
\usage{
create_flight_path(
  deploy_coords,
  land_coords,
  dsm_path,
  N = 100,
  H = 10,
  crs_proj = "EPSG:32632",
  minDist = 1
)
}
\arguments{
\item{deploy_coords}{Numeric vector of length 2 with deployment coordinates (x,y)}

\item{land_coords}{Numeric vector of length 2 with landing coordinates (x,y)}

\item{dsm_path}{Path to Digital Surface Model (DSM) raster file}

\item{N}{Number of sample points along the path (default = 100)}

\item{H}{Desired height above ground level (AGL) in meters (default = 10)}

\item{crs_proj}{Projected CRS (default is UTM Zone 32N - EPSG:32632)}

\item{minDist}{Min distance between trajectory points - UgCS restricted (default = 1)}
}
\value{
A list containing flight path components
}
\description{
Create a smart AGL flight path
}
