% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flight_map.R
\name{plot_flight_map}
\alias{plot_flight_map}
\title{Create interactive map of flight path}
\usage{
plot_flight_map(flight_path)
}
\arguments{
\item{flight_path}{Flight path object created by create_flight_path()}
}
\value{
A leaflet map object
}
\description{
Create interactive map of flight path
}
