% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flight_path.R
\name{plot_flight_path}
\alias{plot_flight_path}
\title{Plot flight path visualization}
\usage{
plot_flight_path(flight_path)
}
\arguments{
\item{flight_path}{Flight path object created by create_flight_path()}
}
\value{
A ggplot object showing the flight path comparison
}
\description{
Plot flight path visualization
}
